refactor: keep deps w/ IdeationComponentWrapper
That is, keep them as file siblings and not in a sub folder
