+++ src/app/(main)/my-voyage/[teamId]/ideation/components/ContributionCard.tsx
