+++ src/app/(main)/my-voyage/[teamId]/ideation/components/FinalizedIdeationCard.tsx
