+++ src/app/(main)/my-voyage/[teamId]/ideation/components/VoteCard.tsx
