+++ src/app/(main)/my-voyage/[teamId]/ideation/components/VoteDescriptionCard.tsx
